backend/pyproject.toml
[project]
name = "audio-streamer-backend"
version = "0.1.0"
description = "Audio Streamer Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=43.0.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "email-validator>=2.2.0",
    "mutagen>=1.47.0",
]
